refactoring, avoid fonction duplication
